framework/requirements.txt
# Magentic Foundation Framework Requirements

# Core Microsoft Agent Framework
agent-framework>=1.0.0

# Azure Integration
azure-identity>=1.15.0
azure-ai-openai>=1.0.0b8
azure-monitor-opentelemetry>=1.2.0

# Web Framework & API
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Research & Search Tools
tavily-python>=0.3.0

# MCP and Protocol Support
websockets>=12.0
aiohttp>=3.9.0
pyyaml>=6.0.1
jsonschema>=4.20.0

# Database & Storage
sqlalchemy>=2.0.25
alembic>=1.13.0
redis>=5.0.1
aiosqlite>=0.19.0

# Observability & Monitoring
opentelemetry-api>=1.22.0
opentelemetry-sdk>=1.22.0
opentelemetry-instrumentation-fastapi>=0.43b0
prometheus-client>=0.19.0
structlog>=23.2.0

# Container & Deployment
docker>=7.0.0
kubernetes>=29.0.0

# Security
cryptography>=41.0.8
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# Async & Concurrency
asyncio-mqtt>=0.16.1
celery>=5.3.4
aiofiles>=23.2.1

# Data Processing
pandas>=2.1.4
numpy>=1.26.2
jinja2>=3.1.2

# Development & Testing
pytest>=7.4.3
pytest-asyncio>=0.21.1
pytest-cov>=4.1.0
black>=23.11.0
isort>=5.12.0
mypy>=1.7.1
pre-commit>=3.6.0

# Documentation
mkdocs>=1.5.3
mkdocs-material>=9.5.0
mkdocstrings[python]>=0.24.0